packages/python/requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
clarifai==10.0.0
clarifai-grpc==10.0.3
click==8.1.7
contextlib2==21.6.0
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.0
frozenlist==1.4.1
fsspec==2023.12.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.0
langchain-community==0.0.12
langchain-core==0.1.10
langsmith==0.0.80
llama-index==0.9.30
markdown-it-py==3.0.0
marshmallow==3.20.2
mdurl==0.1.2
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
openai==1.7.2
opencv-python==4.9.0.80
packaging==23.2
pandas==2.1.4
pillow==10.2.0
protobuf==4.25.2
pydantic==2.5.3
pydantic-core==2.14.6
pygments==2.17.2
pypdf==3.17.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python-rapidjson==1.14
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
schema==0.7.5
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
tenacity==8.2.3
tiktoken==0.5.2
tqdm==4.66.1
tritonclient==2.41.1
typing-extensions==4.9.0
typing-inspect==0.9.0
tzdata==2023.4
urllib3==2.1.0
wrapt==1.16.0
yarl==1.9.4
